+++ klmp3.py
@@ -31,7 +31,7 @@ from tkinter import ttk, filedialog, messagebox
 from ffmpeg_locator import find_ffmpeg_tools_first
 from tab_tags import TagsTab
 from tab_help import HelpTab
-
+from tab_convert import ConvertTab
 
 from PIL import Image, ImageTk
 
@@ -390,7 +390,7 @@ class App(tk.Tk):
         self.ff = find_ffmpeg_tools_first()
         self.ffmpeg_path = self.ff.ffmpeg
         self.ffprobe_path = self.ff.ffprobe
-        self.title("KLMP3 - 2.7.1")
+        self.title("KLMP3 - 2.8")
         self.geometry("820x620")
         self.minsize(780, 560)
 
@@ -514,17 +514,24 @@ class App(tk.Tk):
 
         self.tab_general = ttk.Frame(self.nb)
         self.tab_options = ttk.Frame(self.nb)
+        self.tab_convert = ttk.Frame(self.nb)  
         self.tab_tags = ttk.Frame(self.nb)
         self.tab_help = ttk.Frame(self.nb)
 
         self.nb.add(self.tab_general, text="Général")
         self.nb.add(self.tab_options, text="Options")
+        self.nb.add(self.tab_convert, text="Conversion")  
         self.nb.add(self.tab_tags, text="Métadonnées")
         self.nb.add(self.tab_help, text="Aide")
 
+
+        # Construit l'onglet Conversion (UI + logique) depuis le module séparé
+        self.convert_tab_ui = ConvertTab(parent=self.tab_convert, app=self)
+        self.convert_tab_ui.pack(fill="both", expand=True)
+
         # Construit l'onglet Tags (UI + logique) depuis le module séparé
         self.tags_tab = TagsTab(parent=self.tab_tags, app=self)
-
+   
         # Construit l'onglet Aide (affiche AIDE.md)
         self.help_tab = HelpTab(parent=self.tab_help, md_filename=os.path.join("assets", "AIDE.md"))
         self.help_tab.pack(fill="both", expand=True)